Create HelloWorld.txt
TL;DR
Basically, this is a great starter project that is more of a template.
My two cents are that we can make this project more readable and open. By adding this file to show our inclusiveness, the audience not just feel more at home but also get the motivation to continue the development on this project. What do you think?